# API.yml
openapi: '3.0.2'
info:
  title: API Title
  version: '1.0'
servers:
  - url: http://172.23.2.50:9000
components:
  schemas:
    orderRequestBody:
      type: object
      required:
        - "name"
        - "phoneNumber"
        - "orderDateTime"
        - "numberOfPeople"
        - "numberOfTable"
      properties:
        Name:
          type: string
          example: "Jack"
        Phone:
          type: string
          example: "0912345678"
        DateTime:
          type: string
          example: "2023-12-20 18:00:00"
        NumberOfPeople:
          type: integer
          format: int32
          minimum: 1
          maximum: 12
          example: 5
        Table:
          type: array
          items:
            type: integer
            minimum: 1
            maximum: 36
          example: [1,2,3,4,5]
        Remark:
          type: string
...
paths:
...
  /tables/order:
    post:
      requestBody:
        content:
          'application/json':
            schema:
             $ref: "#/components/schemas/orderRequestBody"
      responses:
        '200':
          description: OK
  /tables/orderStatus:
    get:
      parameters:
        - in: query
          name: tableID
          schema:
            type: integer
      responses:
        '200':
          description: OK
...